Move all previous core/lib modules now in lib/lib up to lib
