+++ src/lib/create-ad-slot.ts
@@ -1,6 +1,6 @@
 import { log } from '@guardian/libs';
 import type { SizeMapping } from './ad-sizes';
-import { isBreakpoint } from './lib/breakpoint';
+import { isBreakpoint } from './breakpoint';
 
 const adSlotIdPrefix = 'dfp-ad--';
